fix: Rename to `ManagerUnifiedConfig`
