+++ src/ai/backend/manager/cli/context.py
@@ -26,7 +26,7 @@ from ai.backend.common.types import RedisConnectionInfo, RedisProfileTarget
 from ai.backend.logging import AbstractLogger, LocalLogger, LogLevel
 from ai.backend.manager.config.bootstrap import BootstrapConfig
 from ai.backend.manager.config.loader.legacy_etcd_loader import LegacyEtcdLoader
-from ai.backend.manager.config.shared import ManagerUnifiedConfig, RedisConfig
+from ai.backend.manager.config.unified import ManagerUnifiedConfig, RedisConfig
 
 
 class CLIContext:
